requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
idna==3.11
lxml==6.0.4
MouseInfo==0.1.3
numpy==2.4.4
pandas==3.0.2
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyMuPDF==1.27.2.2
pypdfium2==5.7.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python3-xlib==0.15
pytweening==1.2.0
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
rank-bm25
python-dotenv
openai
anthropic
google-generativeai